server/requirements.txt
Flask~=3.0.0
graphene~=2.1.9
python-dotenv~=1.0.0
boto3~=1.34.11
